Mostly implement analytics (minus analytics host)
DGAnalytics::GetUserID() is mostly not implemented as it relies on
NKManager and NKSession, which will probably not be added for a while.
DGAnalytics_Redshift is the analytics host on macOS (not sure about
other platforms) and is a very complex class, so it will be its own
commit at some point.